Refactored code structure again
