+++ src/bot/CommandManager.ts
