+++ src/bot/events/clientReady.ts
