+++ src/discord_api_deployment/delete-commands.ts
